Fixed and tested issue 697.  Futures that return `!Send` values are now rejected by a compiler error.
`Sync` and `Send` are now derived by the compiler for `ScopeFuture`, except that the
`Send`-safety of `impl ScopeHandle` is is still asserted unsafety.  I did not review `rayon_core`
to judge whether that is correct.
