+++ rayon-futures/src/compile_fail/mod.rs
@@ -1,2 +1,3 @@
 // These modules contain `compile_fail` doc tests.
 mod future_escape;
+mod non_send_item;
